move and change nats controllers
